backend/requirements.txt
fastapi>=0.115.2,<1.0
uvicorn[standard]==0.27.1
python-multipart>=0.0.18
starlette>=0.40.0,<1.0
jinja2==3.1.3
pydantic[email]>=2.8,<3
httpx==0.27.0
sendgrid>=6.11,<7
pyjwt==2.8.0
cryptography>=42.0,<44
